Adapt code for change in count methode
In php 8.0.0 a change to count(Countable|array) was made and now it throws a type error in this line.
Prior to PHP 8.0.0, if the parameter was neither an array nor an object that implements the Countable interface, 1 would be returned, unless value was null, in which case 0 would be returned.
This change should allow the usage as prior to the change.
Adding  (!is_array($LOGIN) && !(is_countable($LOGIN) would allow the usage as prior to this change, but I am not sure if this is intendend